Keep more logs for Peloton components in vcluster
Summary: Mesos agent on some clusters (such as preprod) are configured to keep only a single log file with a maximum size of 10MB. This is often insufficient to debug problems in Peloton components created as part of vcluster, especially for long performance runs. This change increases the number of additional log files to 10.

Test Plan: Created vcluster, verified that log-rotation config for Peloton components is 10.

Reviewers: O521 Project Peloton: Add reviewers, adityacb

Reviewed By: O521 Project Peloton: Add reviewers, adityacb

Subscribers: jenkins

Tags: #peloton

Maniphest Tasks: T2275717

Differential Revision: https://code.uberinternal.com/D2394773