Use SetPlacment API to return failed placements
Summary:
This diff changes the placement engine and the resource manager to use the `SetPlacement` API
to return the failed placements from the placement engine. Earlier this used to be done via the
`EnqueueGangs` API with the resource pool set to `nil`.

With this change the resource manager will now have the reason for each gang which couldn't be
placed by the placement engine.  The reason is being populated only by the stateless placement
engine for now.

I've also refactored the handler test in the resource manager.

Test Plan:
1. Unit test
2. Tested locally.

Reviewers: O521 Project Peloton: Add reviewers, rcharles, apoorvaj, amitbose

Reviewed By: O521 Project Peloton: Add reviewers, amitbose

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D2399901