Check invalid tasks when dequeueing for placing
Summary:
When a task in `READY` state is killed in the resource manager it needs to be
removed from the ready queue as well otherwise it can be sent to the placement
engine for placement. This doesn't cause any issues in normal operations as the returned
placement will be ignored. But it can cause issues if the task is re-enqueued after
being killed. In that case the placement gang can contain duplicate tasks with the
same peloton task ID.
This will also remove the unnecessary work done by the placement engine for a deleted
task.

The logic to detect an invalid task is done during `Dequeue` since removing from the queue
inplace will be an expensive operation.

Test Plan: Added tests

Reviewers: O521 Project Peloton: Add reviewers, O2101 Project Peloton: Resource Manager, apoorvaj, rcharles

Reviewed By: O521 Project Peloton: Add reviewers, O2101 Project Peloton: Resource Manager, apoorvaj, rcharles

Subscribers: jenkins

Maniphest Tasks: T2403461

Differential Revision: https://code.uberinternal.com/D2416167