Always enqueue ongoing update for stateless jobs
Summary:
This is needed for handling the following scenario
  - Job is in KILLED state
  - A job update is started to remove instances
In the above scenario, once the instances are removed, update is stuck, since the job runtime updater not run.

This diff modifies the behavior of job goalstate engine to always enqueue the ongoing updates into update goalstate engine.

Test Plan: Integration test

Reviewers: zhixin, apoorvaj, O521 Project Peloton: Add reviewers

Reviewed By: zhixin, O521 Project Peloton: Add reviewers

Subscribers: jenkins

Tags: #peloton

Differential Revision: https://code.uberinternal.com/D2420485