Add opaquedata package for simulated unsupported update actions
Summary:
Track update actions in opaque data which are not supported by Peloton:
- startJobUpdate with pulse set
- pulseJobUpdate
- rollbackJobUpdate

Consumers of opaque data can see what the latest action was to determine what state we should
report an update as, e.g. if an update is in a PAUSED state and the latest action was StartPulsed,
then the correct state is ROLLING_FORWARD_AWAITING_PULSE.

Note, I do not expect handler code to be examining actions directly. It must reconcile Peloton
update states with the list of Aurora actions to determine the correct Aurora state.

Correctly implements pulseJobUpdate using opaque data.

Reviewers: varung, apoorvaj, kevinxu, O521 Project Peloton: Add reviewers

Reviewed By: varung, O521 Project Peloton: Add reviewers

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D2430339