Fix state machine errors when transition is a no-op
Summary:
Currently when a statemachine transition is requested and if the statemachine is
already in the desired state we log an error and this create s a lot of spurious logs in
the resource manager. Instead of logging an error this diff checks the error type returned
from the state machine and only logs if it's an error condition.

This diff also removes the exposure of statemachine outside of the `rmTask` and introduces
a `RmTaskState` which has relevant information about the current state of the task instead
of returning the whole statemachine which is bad.

Test Plan: Added tests

Reviewers: O521 Project Peloton: Add reviewers, O2101 Project Peloton: Resource Manager, rcharles

Reviewed By: O521 Project Peloton: Add reviewers, O2101 Project Peloton: Resource Manager, rcharles

Subscribers: sachins, jenkins

Differential Revision: https://code.uberinternal.com/D2440801