Do not set goal state on task initialize
Summary:
Currently, we set the task goal state on task initialization. Due to this there is a data race in the following scenario
  1. Task t of job A(with KillOnPreempt set to false) is preempted as a result of which it is reinitialized.
  2. Job A is stopped via JobStop API which results in killing of all tasks of the job

In the above scenario, the goal state of task t depends on the order in which TaskInitialize and TaskKill are executed
  1. TaskInitialize is executed before TaskKill - GoalState is initially set to RUNNING and then set to KILLED - (EXPECTED BEHAVIOR)
  2. TaskKill is executed before TaskInitialize - GoalState is set to KILLED and then is overwritten to RUNNING - (UNEXPECTED BEHAVIOR)

Since the goalstate is already set to the correct value when TaskInitalize is called, we need not set the goalstate on TaskInitialize. This diff makes this change

Reviewers: zhixin, varung, apoorvaj, O521 Project Peloton: Add reviewers

Reviewed By: varung, O521 Project Peloton: Add reviewers

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D2448157