Re-implement GetJobUpdateDetails / GetJobUpdateSummaries w/ rollback handling
Summary:
If an update is manually rolled back, we generate a new update to revert the original update.
When reading out updates in GetJobUpdateDetails and GetJobUpdateSummaries, we need to detect
update pairs like this and join them together into a single update.

This diff tears out the old implementation of GetJobUpdateDetails to support this use-case.

This diff also completely removes support for JobInstanceUpdateEvents until we figure out
the story with the Peloton GetWorkflowEvents API.

Reviewers: varung, kevinxu, O521 Project Peloton: Add reviewers

Reviewed By: kevinxu, O521 Project Peloton: Add reviewers

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D2482011