Delete stateless job ID from active jobs list on delete
Summary:
For stateless jobs, the job is not untracked even after it
is marked as KILLED. KILLED jobs are still active jobs.
Hence it will not be automatically cleared out of the
active_jobs list when it is KILLED.

Added code to remove the job from active_jobs list as part
of JobDelete action which is the only time a stateless job
will be actually deleted and be no longer active.

Resolves T2454243

Test Plan:
Verified on minicluster.
Added unit tests

Reviewers: O521 Project Peloton: Add reviewers, zhixin

Reviewed By: O521 Project Peloton: Add reviewers, zhixin

Subscribers: apoorvaj, jenkins, sachins

Maniphest Tasks: T2454243

Differential Revision: https://code.uberinternal.com/D2491301