Implement thermos executor in aurora bridge update path.
Summary:
Workflow:
1. When startJobUpdate api is call, aurora bridge extracts `TaskConfig`
   struct from `JobUpdateRequest`.
2. Bridge serializes `TaskConfig` into byte array, and attaches it onto
   `ExecutorInfo.Data`.
3. At the same time, `CommandInfo` and `ExecutorInfo` will be populated
   using flags passed to bridge.

There are some additional changes related with the minicluster. In order for
themros executor working in minicluster, I added a Dockerfile for building
a custom mesos-slave image. Currently the image is manually built, and
pushed to Docker Hub under repo: `kevinxucs/mesos-slave`. This is a bit hacky,
probably needs a better solution.

Test Plan:
Execute unit test. I have not tested it end-to-end, since there is some
additional change needed in order for cli to work with the update path.

Reviewers: varung, codyg, apoorvaj, O521 Project Peloton: Add reviewers

Reviewed By: codyg

Subscribers: min, jenkins

Differential Revision: https://code.uberinternal.com/D2485543