[resource manager] API handler should be blocked until recovery is completed
Summary:
The resource manager runs recovery of running and non-running(pending) tasks during startup. The proper
resource accounting needs to be completed before any requests to mutate the rmtasks is accepted by the
handler. This diff make the resource manager service handler part of the server lifecycle so that it can be
started and stopped with the resource manager gaining and losing leadership. This diff also updates the
startup order of different resource manager components so that:
1. recovery happens before handler is started
2. entitlement calculation is started before task scheduler runs

With this change the  resource manager will not accept any API requests until the recovery is performed
but is necessity to ensure correctness.

Reviewers: O521 Project Peloton: Add reviewers, O2101 Project Peloton: Resource Manager, rcharles, mabansal

Reviewed By: O521 Project Peloton: Add reviewers, O2101 Project Peloton: Resource Manager, mabansal

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D2490897