Change Timer to Histogram for SLA measurement
Summary:
This diff changes the underlying metrics type for measuring SLA's from Summaries(Timers)
to Histograms. There are a couple of reasons for this
1.  It's hard to aggregate Timer's over multiple tags
2. Timer's are more memory intensive than Histograms

Since we can tolerate approximate quantile measurements and need flexible aggregation we can move
to using Histograms instead.

The buckets are chosen such we have enough granularity at lower time intervals and the
last bucket is for approx 1.2 days which should be more than enough.

Reviewers: O521 Project Peloton: Add reviewers, O2101 Project Peloton: Resource Manager, rcharles, chrisriley

Reviewed By: O521 Project Peloton: Add reviewers, O2101 Project Peloton: Resource Manager, rcharles

Subscribers: codyg, seh, jenkins, nick.cobb, chrisriley

Maniphest Tasks: T2348953

Differential Revision: https://code.uberinternal.com/D2508237