Add jobType field to task cache
Summary: Currently, `PatchTaskRuntime` makes a call to the job cache to get the jobType from the jobConfig (write lock if config not present). This means that any function that acquires job lock(read/write) will have to release the lock before calling `PatchTaskRuntime`. Otherwise it's a deadlock. Since jobType is an immutable field, we can add the jobType field to the task cache.

Reviewers: apoorvaj, zhixin, O521 Project Peloton: Add reviewers

Reviewed By: zhixin, O521 Project Peloton: Add reviewers

Subscribers: amitbose, kevinxu, jenkins

Differential Revision: https://code.uberinternal.com/D2520287