Add JobConfigOps to ORM code
Summary:
Add `JobConfigObject` and `JobConfigOps` interface
to storage/objects layer.

Add a `CreateIfNotExist` connector API to implement CAS writes.
(We are still using CAS writes for job config, so have to support this)

The actual migration to this interface will be done in the next diff
and will fix T2387819. Not adding that code here to make the
code review easier.

Resolves T2551803

Test Plan: Added unit tests

Reviewers: O521 Project Peloton: Add reviewers, O2529 Project Peloton: Storage, sishi

Reviewed By: O521 Project Peloton: Add reviewers, O2529 Project Peloton: Storage, sishi

Subscribers: jenkins

Maniphest Tasks: T2387819, T2551803

Differential Revision: https://code.uberinternal.com/D2528525