Set desired host to handle in-place update overwritten
Summary:
If an in-place update is running, the host is reserved
for a task, and a new update overwrites the prev
one, then the new update may get stuck.
It can happen when the old task is in KILLED
state, so in the new update the desired host field
is not set, and the host will be in held state until
after the held state expires and cannot be used to
place other tasks until after the host

Test Plan:
in minicluster, previously in-place update quick overwritten
can cause update stuck.
After the change no stuck is observed.

Reviewers: apoorvaj, varung, O521 Project Peloton: Add reviewers

Reviewed By: apoorvaj, O521 Project Peloton: Add reviewers

Subscribers: jenkins

Maniphest Tasks: T2503699

Differential Revision: https://code.uberinternal.com/D2529859