Fix performance compare script to use production config
Summary:
Peloton apps now us the base64 encoded production.yaml as an environment
variable which gets translated to ./config/${app}/production.yaml inside the container.
When this change was made , vcluster wasn't modified to read the config, encode it and
pass it to the container as the PRODUCTION_CONFIG environ variable. This diff fixes that
by adding another parameter the run-perf-compare.sh script depends on.

Reviewers: O521 Project Peloton: Add reviewers, apoorvaj

Reviewed By: O521 Project Peloton: Add reviewers, apoorvaj

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D2594359