Store task labels in the cache
Summary:
Task labels in the cache will be used to filter the tasks events in the watch API.
Whenever, the configuration version in the task runtime changes,
or tasks get replaced in the cache, update the task labels
in the cache as well. This ensures that the labels stored in
the cache are always pointing to the configuration being
pointed to by the task runtime in the cache.

Reviewers: zhixin, #peloton

Reviewed By: zhixin, #peloton

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D2615673