Support exclusive placement of tasks to host
Summary:
The purpose of this change is to allow Peloton to specific hosts exclusively for certain tasks only. Providing this ability has two aspects:
1. Place tasks on particular hosts only
2. Do not place other tasks on those hosts.

The first part can be achieved through host constraints on a task (i.e. task is only placed on a host that has a matching label). This change implements the second part (anti-affinity) by introducing a special host label called "peloton/exclusive". Such hosts are meant for exclusive use of certain workloads (as indicated by the value of the label). For a task to be placed on such a host, that task must have a host-label constraint with key
"peloton/exclusive" and a value that matches the attribute value. Conversely, if a task doesn't have the exclusive host-label constraint, then all hosts that have the "peloton/exclusive" attribute will be disregarded for placement of that task. This ensures that tasks from other workloads are never placed on hosts with the exclusive attribute.

At this point, the exclusive attribute only works for batch jobs (batch placement plugin). In a subsequent change, the Mimir plugin will be updated so that it will work for stateless/stateful jobs.

Test Plan:
Added UTs and integration tests for the hosts with exclusive attribute. Set up minicluster and verified that hosts marked as exclusive only run tasks that have the
"peloton/exclusive" constraint.

Reviewers: #peloton, avyas

Reviewed By: #peloton, avyas

Subscribers: apoorvaj, varung, jenkins

Tags: #peloton

Maniphest Tasks: T2520613

Differential Revision: https://code.uberinternal.com/D2618577