Implement the logic for host reservation in resource manager to mark tasks ready for host reservation and send tasks to placement.
Summary:
Implements the retry logic in resource manager that each task gets retried for a number of attempts in a cycle and a number of cycles. After that, the task is ready for host reservation. Such task is placed in ready queue and given a larger timeout.

In placement, tasks from resourcemanager are polled and placed into reserveQueue if they are ready for host reservation. Reserver start()/stop() is also added in hostmgr engine.

A configuration enable_host_reservation in resourcemanager can enable/disable host reservation feature.

Test Plan: Tested locally to make sure the tasks get reserved and launched properly.

Reviewers: mabansal, O521 Project Peloton: Add reviewers, zhixin, amitbose

Reviewed By: mabansal, O521 Project Peloton: Add reviewers

Subscribers: jenkins

Maniphest Tasks: T2687653

Differential Revision: https://code.uberinternal.com/D2657913