Add SLASpec to JobSummary
Summary:
SLA spec is a one of the crucial information to qualify for job summary response and it is heavily used by aurorabridge. With addition of SLA spec in job summary, it will prevent calling jobInfo which is much heavier call w.r.t. to job summary.

Changeset includes
- Add sla spec in job_index table, but it is not required to index.
- Update job summary response to include sla spec, and update read/write path to include the information.
- Changes at aurora bridge to use job summary rather than job info.

Reviewers: apoorvaj, kevinxu, #peloton

Reviewed By: kevinxu, #peloton

Subscribers: adityacb, jenkins

Maniphest Tasks: T2677531

Differential Revision: https://code.uberinternal.com/D2649799