Fix job level stop and pod level start conflict
Summary:
When using workflow to stop a job (which is used in
v0 API), the workflow may get stuck when pod starts
via pod API. The root cause is that the workflow
would think that the instance is in the progress of
kill, and wait for it to be killed, but it would never
be. This diff handles this case, and treats the pod stop
as completed if it is later started by user again.

Reviewers: sachins, apoorvaj, #peloton

Reviewed By: apoorvaj, #peloton

Subscribers: jenkins

Maniphest Tasks: T2278401

Differential Revision: https://code.uberinternal.com/D2664755