Fix ORM panic when reading a column that is null
Summary:
There was a bug in ORM code at the location where we translate
the row read from the DB connector to individual object fields.
In case where the DB has value for a particular column as null,
we should check the reflected valueOf for validity and also check
if it is a pointer to Nil.

The earlier check of making sure that the columnValue is nil is
not good enough. Because in this case it will contain (*type).
Added another check as well as a unit test.

Resolves T2705003

Reviewers: O521 Project Peloton: Add reviewers, amitbose

Reviewed By: O521 Project Peloton: Add reviewers, amitbose

Subscribers: varung, jenkins

Maniphest Tasks: T2705003

Differential Revision: https://code.uberinternal.com/D2674505