Enqueue job to job goal state engine if job state is already KILLED
Summary:
In a scenario of zero instances are created in a job (with config instance count >0). If a job is KILLED and then DELETED, then it is not enqueued in job goal state engine since task stats count match.

Add a check to enqueue job to goal state engine if job is already KILLED.

Test Plan: Integ Test

Reviewers: apoorvaj, zhixin, O521 Project Peloton: Add reviewers

Reviewed By: zhixin, O521 Project Peloton: Add reviewers

Subscribers: jenkins

Maniphest Tasks: T2693545

Differential Revision: https://code.uberinternal.com/D2674907