Mimir placement plugin host filtering
Summary:
Host filtering based on constraints now happens in host-manager
for the Mimir placement plugin as well (this was already the
case for batch plugin). This allows for potentially faster
placement when there are (host) constraints and support for
exclusive hosts in Mimir.

Test Plan:
Updated Mimir unit-tests. Added integration tests for
placement of stateless jobs with and without exclusive
host constraints.

Reviewers: #peloton, avyas, varung, O521 Project Peloton: Add reviewers

Reviewed By: #peloton, avyas, O521 Project Peloton: Add reviewers

Subscribers: jenkins

Maniphest Tasks: T2687525, T2520613

Differential Revision: https://code.uberinternal.com/D2680959