Fix job query for partially created job
Summary:
If job_index does not contain a valid name, job query will construct
the job summary from job_config. This was put in when "name" was
added to the job_index table, to handle the older job_index entries
for whom, name is "".

Corner case:
In case a job which genuinely has no name, is partially created, the
construction of job summary from job_config will fail because config
does not exist which happens in the reported bug.

Added a fix to not throw error and move on to next entry in case the
query to job_config fails for any reason.

Reviewers: O521 Project Peloton: Add reviewers, sachins

Reviewed By: O521 Project Peloton: Add reviewers, sachins

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D2682521