Introduce a mini flag in perf test to run just one test
Summary:
Current version of perf test takes too long.
We know that one test (50k tasks, no instance config, sleep=10s)
is the one we care about the most when debugging perf issues.
Historically, a regression in numbers for this test indicates
something is wrong in code.

Added a `--mini` flag to the perf test suite to just run this one
test. Will make this configurable from Jenkins.

This reduces test times and encourages folks to debug perf issues
as and when they see it.

Reviewers: O521 Project Peloton: Add reviewers, rcharles

Reviewed By: O521 Project Peloton: Add reviewers, rcharles

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D2692911