Specify InvalidEntityVersionError to be part of API
Summary:
InvalidEntityVersionError and UnexpectedVersionError are
special errors that require users to have a specific action.
In more details, the error means there is a write-after-reaed conflict,
and multiple users are trying to operate on the job at the same time.
Therefore  users need to read the job again and get the latest job status,
if users reeceive the error.

Users need to get the error code and error message to know they
get this particualr error, so the error is part of API. Unfortunately,
neither yarpc v1 nor proto supports error as part of API.

This diff adds comments and tests to remind developers any change
to the errors under the file can be a breaking change. Also,
it changes the error type of InvalidEntityVersionError, because,
InvalidEntityVersionError and UnexpectedVersionError are both due to
the same kind of concurrency issue and both should share the same error
code.

Reviewers: apoorvaj, O521 Project Peloton: Add reviewers

Reviewed By: apoorvaj, O521 Project Peloton: Add reviewers

Subscribers: kevinxu, jenkins

Maniphest Tasks: T2342747

Differential Revision: https://code.uberinternal.com/D2699399