Add ORM method for iterative get-all
Summary:
ORM Client and Connector interfaces will now support
an iterative version of the GetAll method - the
caller gets an Iterator object on which Next() must
be called repeatedly to get the rows in the result-set.
Close() must be called when the iterator is no longer
needed. When the result-set of a query is large,
iterative read enables you to avoid buffering all the
results in memory.
Note: I moved the orm test code to package orm_test
to workaround circular dependency issues with mocks.

Test Plan: Unit tests and integration tests.

Reviewers: O521 Project Peloton: Add reviewers, adityacb, sishi

Reviewed By: O521 Project Peloton: Add reviewers, adityacb

Subscribers: varung, jenkins

Maniphest Tasks: T2696885

Differential Revision: https://code.uberinternal.com/D2723315