Reduce host pruning period from 10m to 2m
Summary:
In current state, host pruner runs at 10 minutes and host expiration is 5 minutes. So in total a host stay in placing state for 15 minutes.

This diff reduces the host pruner to run every 2 minutes, and host expiration is 5 minutes. So a max host stays in placing state is 7 minutes.

Reviewers: apoorvaj, zhixin, kevinxu, O521 Project Peloton: Add reviewers, sachins

Reviewed By: O521 Project Peloton: Add reviewers, sachins

Subscribers: jenkins

Maniphest Tasks: T2746849

Differential Revision: https://code.uberinternal.com/D2742569