Remove GetLastRuntimeUpdateTime for timeout
Summary:
Peloton uses cachedTask.GetLastRuntimeUpdateTime to
check if a task action timeouts and needs to be retried.
cachedTask serves the request from cache, which means
upon restart it would return 0 and all of the action
checked would be considered to have timeout.

Use taskRuntime.Revision.UpdateAt instead. This field
is persisted in DB, and is persistent between restarts.

Reviewers: O521 Project Peloton: Add reviewers, sachins, apoorvaj

Reviewed By: O521 Project Peloton: Add reviewers, apoorvaj

Subscribers: jenkins

Maniphest Tasks: T2809335

Differential Revision: https://code.uberinternal.com/D2753385