Create a new context to cleanup previous job updates
Summary:
When a new job update is created, a new go routine is created
to cleanup any old job updates in the storage. However, the same
context coming in from yarpc is used in the go routine. Since the go
routine can keep running after the rpc call returns, this causes the
context to get cancelled and the updates are not cleaned up.
This change creates a new context for the go routine cleaning
up the older job updates.

Reviewers: O521 Project Peloton: Add reviewers, zhixin

Reviewed By: O521 Project Peloton: Add reviewers, zhixin

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D2772709