ListJobs API should not fail when update-fetch from store fails for one of the jobs
Summary:
ListJobs does the following
1. Fetch job summaries for the all jobs in job index.
2. For each job, fetch update information and send it to the stream.

Earlier, while performing step 2, if it runs into an error, it immediately bails out and returns an error. Now, on error while fetching a job update, it will populate the response with whatever is available and move-on to the next job.

Test Plan: - Modified unit tests

Reviewers: apoorvaj, O521 Project Peloton: Add reviewers

Reviewed By: apoorvaj, O521 Project Peloton: Add reviewers

Subscribers: jenkins

Maniphest Tasks: T2892037

Differential Revision: https://code.uberinternal.com/D2789343