Change update state in DB only if update state changes
Summary:
Currently updateStore.WriteProgress is called each time
when UpdateRun is called, no matter whether there is real
change or not. This would change the last update time field
in the update table, so even if nothing changes, we may still
infer the update state changes.

This diff solves the issue by only write to DB when the update
status does change.

Test Plan:
tested on pcluster. With some additional log, the test observed
write to DB when nothing really changes when running old code.
When running new code, this is not observed.

Reviewers: apoorvaj, sachins, O521 Project Peloton: Add reviewers

Reviewed By: sachins, O521 Project Peloton: Add reviewers

Subscribers: jenkins

Maniphest Tasks: T2900517

Differential Revision: https://code.uberinternal.com/D2795033