Fix resource leak issues in resource manager
Summary:
This diff fixes the below 2 issues
1. Resources held by orphan task, not present in tracker
Before, when resmgr receives a Mesos event for a task, a check is made to see if the task is present in the tracker. If it is not, then the event is dropped. This can cause issues in cases where the event for the orphan task arrives after the task is removed from the tracker (instance count reduction in case of stateless tasks and task completion in case of batch) causing the resources to be held indefinitely (until resmgr is restarted).

This diff cleans up for the orphan task even if the task is not present in the tracker.

2. Resources held by orphan tasks in READY queue
In Job Manager, when the task launch fails (LAUNCHING state for the associated RMTask), we requeue the task to Resource Manager without killing the previous RM task causing the previous RMTask to become an orphan RMtask. The resources held up by orphan RMTask are released only on receiving a Mesos terminal event for the task. Since this orphan task doesnt have an associated Mesos task, these resources will not be released until we restart the resmgr.

This diff makes the change to allow requeuing of tasks only in LAUNCHED and RUNNING state (and not LAUNCHING)

Test Plan:
- Added unit tests
- Added integration tests. Integration tests (with large instance counts) were tested on wbu2-da2
- Tested on pit1-preprod01 and wbu2-da2 by creating a job with instance count 9000 (0.1CPU per instance).

Reviewers: avyas, varung, O521 Project Peloton: Add reviewers, zhixin, apoorvaj

Reviewed By: avyas, O521 Project Peloton: Add reviewers, apoorvaj

Subscribers: jenkins, apoorvaj

Tags: #peloton

Differential Revision: https://code.uberinternal.com/D2780925