Fix corner cases where respool creation crashes resmgr
Summary:
tested 4 scenarios that crashes resmgr
--when resources are not specified in the respool config
--when resources are specified with negative numbers
--when resources have repetitive types
--when totalChildShares are 0

This fix is to update respoolconfig_validator to check on the above scenarios.
This fix also adds default values (reservation:0, limit:0, share:1) if a type of resource is not configured.
This fix also handles when totalChildShares==0 in entitlement calculator

Test Plan: tested locally with minicluster

Reviewers: mabansal, avyas, O521 Project Peloton: Add reviewers, binz

Reviewed By: avyas, O521 Project Peloton: Add reviewers

Subscribers: jenkins, zhixin

Differential Revision: https://code.uberinternal.com/D2801709