Disable starting state timeout for service job
Summary:
Service jobs may be using custom excutor such as thermos.
Thermos would stuck in Starting state and transit to RUNNING,
only after the task passes thermos task health check.
Therefore, it is very likely to for a task to timeout in
STARTING state if peloton is using thermos.

Disbale the timeout check for service job for now, until
we have a better way to handle the issue.

Reviewers: sachins, apoorvaj, O521 Project Peloton: Add reviewers

Reviewed By: apoorvaj, O521 Project Peloton: Add reviewers

Subscribers: jenkins

Maniphest Tasks: T2835627

Differential Revision: https://code.uberinternal.com/D2830961