Placement hints for batch jobs
Summary:
In the batch placement plugin, we try to pack as
many tasks on a host as possible, which can have
adverse effects on task performance due to
interference (for example, if all tasks require
high disk I/O). Such jobs can can work better if
the tasks are spread out across many hosts.
This change adds the notion of a placement hint
to batch jobs. Two strategies are supported:
* Pack: the current behavior
* Spread: distribute tasks over many hosts
The spread strategy doesn't work very well when
defragmentation is enabled for bin-packing. So
defragmentation is explicitly disabled for such
jobs.
For jobs that use a spread strategy, we calculate
"spread quotient" which is defined as the ratio of
number of tasks to number of unique hosts. The
average of spread quotients is tracked to measure
the effectiveness of spreading - values closer to
1 are desirable.

Test Plan:
On minicluster, created jobs with default (pack) and
spread placement hints. Verified that tasks very
placed as expected in both cases.
Enabled defragmentation, and verified that jobs with
spread hint are placed as expected.

Reviewers: O521 Project Peloton: Add reviewers, mabansal, avyas, varung

Reviewed By: O521 Project Peloton: Add reviewers, mabansal

Subscribers: apoorvaj, jenkins

Maniphest Tasks: T2938775

Differential Revision: https://code.uberinternal.com/D2847817