Always retry terminated pod no matter workflow fails or not
Summary:
In the current code, failed pod would stop retry if it exceeds
max failure config in updateConfig when the pod is in the process
of update.

However, the pod would begin to retry infinitely if it is not in
the process of an update.

For example, a job with 3 pods is being updated. pod0 is considered to
have update completed, and pod01 is considered being actively updated.
If both have failed more than max failure allowed in the active update,
pod0 would still get retried bu pod01 would stop in the failed state.

This may result in confusing behavior to user. The diff makes sure a
failed pod would always get retried with throttling.

Reviewers: apoorvaj, sachins, O521 Project Peloton: Add reviewers

Reviewed By: sachins, O521 Project Peloton: Add reviewers

Subscribers: jenkins

Maniphest Tasks: T2822309

Differential Revision: https://code.uberinternal.com/D2845955