[placement] Make placement period configurable
Summary:
This diff makes the placement period configurable along with the dequeue limit. This
will allow controlling the placement rate by setting the appropriate values for the period and
the limit.

For eg: if the `task_dequeue_period`=60s and `task_dequeue_limit`=100 and there are `3` placement
engines then 3*60*100=1800 tasks would be dequeued from the resource manager in 1 minute. This is
usefull when a predictable rate of placements is desired.

Test Plan: Tested locally via minicluster

Reviewers: O521 Project Peloton: Add reviewers, amitbose, varung, adityacb

Reviewed By: O521 Project Peloton: Add reviewers, amitbose, adityacb

Subscribers: adityacb, jenkins

Maniphest Tasks: T3049969

Differential Revision: https://code.uberinternal.com/D2871883