Implement getJobUpdateDetails/Summary using QueryJobCache private api
Summary:
When a deployment happens in udeploy, the workflow happens in the following
order:

1. aggregator calls startJobUpdate with "wait for pulse" enabled
2. aggregator calls getJobUpdateDetails (which uses QueryJobs) to get list
   of updates that requires "pulse"
3. aggregator calls pulseJobUpdate for updates returned by previous step

However, if a new job is created by startJobUpdate call, QueryJobs needs to
wait for lucene index to build so that it will return the corresponding jobs.

Change to use jobmgr's private API - QueryJobCache to query job id and name
based on the label.

Reviewers: varung, zhixin, apoorvaj, O521 Project Peloton: Add reviewers

Reviewed By: zhixin, O521 Project Peloton: Add reviewers

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D2874869