Fix job cache query nil pointer
Summary:
It is possible that job config is deleted, but the job
is still in jobFactory. As a result, job config cannot
be found upon read. To fix the issue, peloton has to
make sure  cached job is removed from jobFactory when
job is deleted. Before the fix is done, ignore a job
if config is not found.

Reviewers: sachins, kevinxu, O521 Project Peloton: Add reviewers

Reviewed By: sachins, kevinxu, O521 Project Peloton: Add reviewers

Subscribers: jenkins

Maniphest Tasks: T3082637

Differential Revision: https://code.uberinternal.com/D2887651