Task is not getting reserved when max_duration in placement is less than placement timeout in resmgr
Summary:
Currently resmgr checks if the task should be ready for host reservation when placing times out. If max_duration in placement is less than placement timeout in resmgr, the task could timeout in placement while waiting for resource and moves into ready state, and placing state of the task never hits the timeout.

This patch now checks host reservation when the task is getting placed in placement. Thus, regardless of the configuration, if all the placement cycles are exhausted, it should be ready for host reservation.

Test Plan: unit test, integration test

Reviewers: O521 Project Peloton: Add reviewers, mabansal, avyas

Reviewed By: O521 Project Peloton: Add reviewers, mabansal

Subscribers: jenkins

Maniphest Tasks: T3091129

Differential Revision: https://code.uberinternal.com/D2895323