Changed placement strategy plugin interface
Summary:
We used to rely heavily on the input types and do in-placement mutation
of those inputs when we passed them into the placement strategy plugin.
This patch changes the `PlaceOnce` function to not do in-place mutation
of those objects but rather returns the placement map `map[int]int`. This
essentially maps the tasks array onto the hosts array through the
indices of their elements.
i.e: `placements[x] == y` means that `tasks[x]` was placed on `hosts[y]`.
Additionally: If `tasks[i]` cannot get placed, `PlaceOnce(tasks, hosts)[i] == -1`.

Note: PlaceOnce was renamed to GetTaskPlacements due to the fact that it no
longer does in-place mutations of the inputs, but rather returns the placement
information to the caller.

This change's goal is to continue to make the placement strategy logic
independent of the API structs and their internals so that we may work
with the v1alpha API soon.

Reviewers: adityacb, O521 Project Peloton: Add reviewers

Reviewed By: adityacb, O521 Project Peloton: Add reviewers

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D2913201