mimir strategy filter group assignments by resource constraint
Summary:
In the current code, mimir filter create batches by grouping tasks with same
resource usage together and takes the max resources as the constraint.

It means that if a big task is scheduled with small tasks, small tasks may not
be able to get placed. This is particularly an issue for in-place update, because
the task tries to get placed on the previous host until number of retries
gets exhausted. And host will be in reserved state until placement engine gives up to
place the task on the host it previous runs on.
Also, for in-place update tasks, the host may just have the exact
amount of resources needed to run the tasks, so max resource constraint does not
work well in the case.

Reviewers: varung, adityacb, pourchet, apoorvaj, O521 Project Peloton: Add reviewers, amitbose

Reviewed By: pourchet, O521 Project Peloton: Add reviewers, amitbose

Subscribers: jenkins

Maniphest Tasks: T2473493

Differential Revision: https://code.uberinternal.com/D2936627